requirements.txt
flask
peewee
pymysql
python-dotenv
flask_wtf